Revert "Ajuna Polkadot parachain (#38)"
This reverts commit c897d90a6ed183c0475666713b4207ccea4dd5e0.
