+++ .github/workflows/check-pull-request.yml
@@ -82,7 +82,7 @@ jobs:
       - uses: actions-rs/cargo@v1
         with:
           command: test
-          args: --workspace --all-features ${{ env.CARGO_ARGS }} --jobs 1
+          args: --workspace --all-features ${{ env.CARGO_ARGS }}
 
   notify:
     name: Notify on Slack
