+++ Cargo.toml
@@ -9,7 +9,6 @@ members = [
     "rpc",
     "runtime/solo",
     "runtime/bajun",
-    "runtime/ajuna",
     "pallets/ajuna-matchmaker",
     "pallets/ajuna-gameregistry",
     "pallets/ajuna-runner",
