--- node/cli/src/command/bajun.rs
@@ -1,259 +1,0 @@
-// Ajuna Node
-// Copyright (C) 2022 BlogaTech AG
-
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU Affero General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU Affero General Public License for more details.
-
-// You should have received a copy of the GNU Affero General Public License
-// along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
-use crate::cli::RelayChainCli;
-use codec::Encode;
-use cumulus_client_service::genesis::generate_genesis_block;
-use cumulus_primitives_core::ParaId;
-use log::info;
-use polkadot_parachain::primitives::AccountIdConversion;
-use sc_cli::Result;
-use sp_core::hexdisplay::HexDisplay;
-use sp_runtime::traits::Block as BlockT;
-use std::io::Write;
-
-use crate::cli::{Cli, Subcommand};
-use ajuna_service::{
-	ajuna_solo_runtime::Block as SoloBlock,
-	bajun_runtime::{Block as ParaBlock, RuntimeApi},
-	chain_spec, para_bajun as para,
-	para_bajun::BajunRuntimeExecutor,
-	solo::ExecutorDispatch,
-};
-use frame_benchmarking_cli::BenchmarkCmd;
-use sc_cli::SubstrateCli;
-use std::path::PathBuf;
-
-fn load_spec(id: &str) -> std::result::Result<Box<dyn sc_service::ChainSpec>, String> {
-	Ok(match id {
-		"dev" => Box::new(chain_spec::bajun::development_config()),
-		"" | "local" => Box::new(chain_spec::bajun::local_testnet_config()),
-		path => Box::new(chain_spec::bajun::ChainSpec::from_json_file(PathBuf::from(path))?),
-	})
-}
-
-#[allow(clippy::borrowed_box)]
-fn extract_genesis_wasm(chain_spec: &Box<dyn sc_service::ChainSpec>) -> Result<Vec<u8>> {
-	let mut storage = chain_spec.build_storage()?;
-
-	storage
-		.top
-		.remove(sp_core::storage::well_known_keys::CODE)
-		.ok_or_else(|| "Could not find wasm file in genesis state!".into())
-}
-
-macro_rules! construct_async_run {
-	(|$components:ident, $cli:ident, $cmd:ident, $config:ident| $( $code:tt )* ) => {{
-		let runner = $cli.create_runner($cmd)?;
-		runner.async_run(|$config| {
-            let $components = para::new_partial::<RuntimeApi, BajunRuntimeExecutor, _>(
-                &$config,
-                para::parachain_build_import_queue,
-            )?;
-
-			let task_manager = $components.task_manager;
-            { $( $code )* }.map(|v| (v, task_manager))
-		})
-	}}
-}
-
-macro_rules! construct_sync_run {
-	(|$components:ident, $cli:ident, $cmd:ident, $config:ident| $( $code:tt )* ) => {{
-		let runner = $cli.create_runner($cmd)?;
-		runner.sync_run(|$config| {
-            let $components = para::new_partial::<RuntimeApi, BajunRuntimeExecutor, _>(
-                &$config,
-                para::parachain_build_import_queue,
-            )?;
-
-            { $( $code )* }
-		})
-	}}
-}
-
-/// Parse and run command line arguments
-#[allow(unused_variables)]
-pub fn run() -> sc_cli::Result<()> {
-	let cli = Cli::from_args();
-
-	match &cli.subcommand {
-		Some(Subcommand::Key(cmd)) => cmd.run(&cli),
-		Some(Subcommand::ExportGenesisState(params)) => {
-			let mut builder = sc_cli::LoggerBuilder::new("");
-			builder.with_profiling(sc_tracing::TracingReceiver::Log, "");
-			let _ = builder.init();
-
-			let spec = load_spec(&params.chain.clone().unwrap_or_default())?;
-			let state_version = Cli::native_runtime_version(&spec).state_version();
-			let block: ParaBlock = generate_genesis_block(&spec, state_version)?;
-			let raw_header = block.header().encode();
-			let output_buf = if params.raw {
-				raw_header
-			} else {
-				format!("0x{:?}", HexDisplay::from(&block.header().encode())).into_bytes()
-			};
-
-			if let Some(output) = &params.output {
-				std::fs::write(output, output_buf)?;
-			} else {
-				std::io::stdout().write_all(&output_buf)?;
-			}
-
-			Ok(())
-		},
-		Some(Subcommand::ExportGenesisWasm(params)) => {
-			let mut builder = sc_cli::LoggerBuilder::new("");
-			builder.with_profiling(sc_tracing::TracingReceiver::Log, "");
-			let _ = builder.init();
-
-			let raw_wasm_blob =
-				extract_genesis_wasm(&cli.load_spec(&params.chain.clone().unwrap_or_default())?)?;
-			let output_buf = if params.raw {
-				raw_wasm_blob
-			} else {
-				format!("0x{:?}", HexDisplay::from(&raw_wasm_blob)).into_bytes()
-			};
-
-			if let Some(output) = &params.output {
-				std::fs::write(output, output_buf)?;
-			} else {
-				std::io::stdout().write_all(&output_buf)?;
-			}
-
-			Ok(())
-		},
-		Some(Subcommand::BuildSpec(cmd)) => {
-			let runner = cli.create_runner(cmd)?;
-			runner.sync_run(|config| cmd.run(config.chain_spec, config.network))
-		},
-		Some(Subcommand::CheckBlock(cmd)) => {
-			construct_async_run!(|components, cli, cmd, config| {
-				Ok(cmd.run(components.client, components.import_queue))
-			})
-		},
-		Some(Subcommand::ExportBlocks(cmd)) => {
-			construct_async_run!(|components, cli, cmd, config| {
-				Ok(cmd.run(components.client, config.database))
-			})
-		},
-		Some(Subcommand::ExportState(cmd)) => {
-			construct_async_run!(|components, cli, cmd, config| {
-				Ok(cmd.run(components.client, config.chain_spec))
-			})
-		},
-		Some(Subcommand::ImportBlocks(cmd)) => {
-			construct_async_run!(|components, cli, cmd, config| {
-				Ok(cmd.run(components.client, components.import_queue))
-			})
-		},
-		Some(Subcommand::PurgeChainPara(cmd)) => {
-			let runner = cli.create_runner(cmd)?;
-
-			runner.sync_run(|config| {
-				let polkadot_cli = RelayChainCli::new(
-					&config,
-					[RelayChainCli::executable_name()].iter().chain(cli.relay_chain_args.iter()),
-				);
-
-				let polkadot_config = SubstrateCli::create_configuration(
-					&polkadot_cli,
-					&polkadot_cli,
-					config.tokio_handle.clone(),
-				)
-				.map_err(|err| format!("Relay chain argument error: {}", err))?;
-
-				cmd.run(config, polkadot_config)
-			})
-		},
-		Some(Subcommand::Revert(cmd)) =>
-			return construct_async_run!(|components, cli, cmd, config| {
-				Ok(cmd.run(components.client, components.backend, None))
-			}),
-		Some(Subcommand::Benchmark(cmd)) => {
-			let runner = cli.create_runner(cmd)?;
-			// Switch on the concrete benchmark sub-command-
-			match cmd {
-				BenchmarkCmd::Pallet(cmd) => runner.sync_run(|config| {
-					if cfg!(feature = "runtime-benchmarks") {
-						cmd.run::<SoloBlock, ExecutorDispatch>(config)
-					} else {
-						Err("Runtime benchmarking wasn't enabled when building the node. \
-                            You can enable it with `--features runtime-benchmarks`."
-							.into())
-					}
-				}),
-				BenchmarkCmd::Block(cmd) => {
-					construct_sync_run!(|components, cli, cmd, config| {
-						cmd.run(components.client)
-					})
-				},
-				BenchmarkCmd::Storage(cmd) => {
-					construct_sync_run!(|components, cli, cmd, config| {
-						let db = components.backend.expose_db();
-						let storage = components.backend.expose_storage();
-						cmd.run(config, components.client, db, storage)
-					})
-				},
-				BenchmarkCmd::Overhead(cmd) => Err("Unsupported benchmarking command".into()),
-				BenchmarkCmd::Machine(cmd) => {
-					construct_sync_run!(|components, cli, cmd, config| cmd.run(&config))
-				},
-			}
-		},
-		Some(_) => Err("Unsupported command for bajun parachain".into()),
-		None => {
-			let runner = cli.create_runner(&cli.run_para.normalize())?;
-			let collator_options = cli.run_para.collator_options();
-
-			runner.run_node_until_exit(|config| async move {
-				let para_id = chain_spec::Extensions::try_get(&*config.chain_spec)
-					.map(|e| e.para_id)
-					.ok_or("Could not find parachain ID in chain-spec.")?;
-
-				let polkadot_cli = RelayChainCli::new(
-					&config,
-					[RelayChainCli::executable_name()].iter().chain(cli.relay_chain_args.iter()),
-				);
-
-				let id = ParaId::from(para_id);
-
-				let parachain_account =
-					AccountIdConversion::<polkadot_primitives::v2::AccountId>::into_account(&id);
-
-				let state_version =
-					RelayChainCli::native_runtime_version(&config.chain_spec).state_version();
-				let block: ParaBlock = generate_genesis_block(&config.chain_spec, state_version)
-					.map_err(|e| format!("{:?}", e))?;
-				let genesis_state = format!("0x{:?}", HexDisplay::from(&block.header().encode()));
-
-				let tokio_handle = config.tokio_handle.clone();
-				let polkadot_config =
-					SubstrateCli::create_configuration(&polkadot_cli, &polkadot_cli, tokio_handle)
-						.map_err(|err| format!("Relay chain argument error: {}", err))?;
-
-				info!("Parachain id: {:?}", id);
-				info!("Parachain Account: {}", parachain_account);
-				info!("Parachain genesis state: {}", genesis_state);
-				info!("Is collating: {}", if config.role.is_authority() { "yes" } else { "no" });
-
-				para::start_parachain_node(config, polkadot_config, collator_options, id)
-					.await
-					.map(|r| r.0)
-					.map_err(Into::into)
-			})
-		},
-	}
-}
