--- node/cli/src/command/mod.rs
@@ -1,8 +1,0 @@
-#[cfg(feature = "solo")]
-pub mod solo;
-
-#[cfg(feature = "bajun")]
-pub mod bajun;
-
-#[cfg(feature = "ajuna")]
-pub mod ajuna;
