+++ node/cli/src/lib.rs
@@ -16,7 +16,6 @@
 
 mod cli;
 mod command;
-
 mod command_helper;
 
 pub use cli::*;
