+++ node/service/Cargo.toml
@@ -7,25 +7,24 @@ repository = "https://github.com/ajuna-network/Ajuna"
 version    = "0.1.0"
 
 [dependencies]
-hex-literal = "0.3.3"
+hex-literal  = "0.3.3"
 jsonrpc-core = "18.0.0"
-serde = { version = "1.0.132", features = ["derive"] }
+serde        = { version = "1.0.132", features = [ "derive" ] }
 
 # Local
-ajuna-primitives = { path = "../../primitives" }
-ajuna-rpc = { path = "../../rpc" }
+ajuna-primitives   = { path = "../../primitives" }
+ajuna-rpc          = { path = "../../rpc" }
 ajuna-solo-runtime = { path = "../../runtime/solo", optional = true }
-bajun-runtime = { path = "../../runtime/bajun", optional = true }
-ajuna-runtime = { path = "../../runtime/ajuna", optional = true }
+bajun-runtime      = { path = "../../runtime/bajun", optional = true }
 
 # Substrate
-frame-benchmarking = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.20" }
+frame-benchmarking  = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.20" }
 sc-basic-authorship = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.20" }
-sc-chain-spec = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.20" }
-sc-client-api = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.20" }
-sc-consensus = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.20" }
-sc-consensus-aura = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.20" }
-sc-executor = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.20", features = ["wasmtime"] }
+sc-chain-spec       = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.20" }
+sc-client-api       = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.20" }
+sc-consensus        = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.20" }
+sc-consensus-aura   = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.20" }
+sc-executor         = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.20", features = [ "wasmtime" ] }
 sc-finality-grandpa = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.20" }
 sc-keystore         = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.20" }
 sc-service          = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.20", features = [ "wasmtime" ] }
@@ -68,35 +67,7 @@ cumulus-relay-chain-interface           = { git = "https://github.com/paritytech
 cumulus-relay-chain-rpc-interface       = { git = "https://github.com/paritytech/cumulus", branch = "polkadot-v0.9.20", optional = true }
 
 [features]
-default = ["solo"]
-
-ajuna = [
-    "ajuna-runtime",
-    "pallet-transaction-payment-rpc",
-    "sc-network",
-    "sc-rpc",
-    "sc-tracing",
-    "sp-api",
-    "sp-block-builder",
-    "sp-keystore",
-    "sp-offchain",
-    "sp-session",
-    "sp-transaction-pool",
-    "substrate-frame-rpc-system",
-    "substrate-prometheus-endpoint",
-    "polkadot-service",
-    "xcm",
-    "cumulus-client-cli",
-    "cumulus-client-consensus-aura",
-    "cumulus-client-consensus-common",
-    "cumulus-client-network",
-    "cumulus-client-service",
-    "cumulus-primitives-core",
-    "cumulus-primitives-parachain-inherent",
-    "cumulus-relay-chain-inprocess-interface",
-    "cumulus-relay-chain-interface",
-    "cumulus-relay-chain-rpc-interface",
-]
+default = [ "solo" ]
 
 bajun = [
     "bajun-runtime",
@@ -125,17 +96,14 @@ bajun = [
     "cumulus-relay-chain-interface",
     "cumulus-relay-chain-rpc-interface",
 ]
+solo = [ "ajuna-solo-runtime" ]
 
-solo = ["ajuna-solo-runtime"]
-
-polkadot-native = ["polkadot-service/polkadot-native"]
-kusama-native = ["polkadot-service/kusama-native"]
-rococo-native = ["polkadot-service/rococo-native"]
+kusama-native = [ "polkadot-service/kusama-native" ]
+rococo-native = [ "polkadot-service/rococo-native" ]
 
 runtime-benchmarks = [
     "ajuna-solo-runtime/runtime-benchmarks",
     "bajun-runtime/runtime-benchmarks",
-    "ajuna-runtime/runtime-benchmarks",
 ]
 
-skip-ias-check = ["ajuna-solo-runtime/skip-ias-check"]
+skip-ias-check = [ "ajuna-solo-runtime/skip-ias-check" ]
