+++ node/service/src/lib.rs
@@ -16,15 +16,11 @@
 
 pub mod chain_spec;
 
-#[cfg(feature = "ajuna")]
-pub mod para_ajuna;
 #[cfg(feature = "bajun")]
-pub mod para_bajun;
+pub mod para;
 #[cfg(feature = "solo")]
 pub mod solo;
 
-#[cfg(feature = "ajuna")]
-pub use ajuna_runtime;
 #[cfg(feature = "solo")]
 pub use ajuna_solo_runtime;
 #[cfg(feature = "bajun")]
