--- resources/polkadot/genesis-state
@@ -1,1 +1,0 @@
-0x0000000000000000000000000000000000000000000000000000000000000000006e666a1df855628a99876f9f876b94d6d20a397fb1b5d92a3747df75e29c61b103170a2e7597b7b7e3d84c05391d139a62b157e78786d8c082f29dcf4c11131400
