--- runtime/ajuna/src/weights/mod.rs
@@ -1,9 +1,0 @@
-pub mod block_weights;
-pub mod extrinsic_weights;
-pub mod paritydb_weights;
-pub mod rocksdb_weights;
-
-pub use block_weights::constants::BlockExecutionWeight;
-pub use extrinsic_weights::constants::ExtrinsicBaseWeight;
-pub use paritydb_weights::constants::ParityDbWeight;
-pub use rocksdb_weights::constants::RocksDbWeight;
