+++ rust-toolchain.toml
@@ -1,5 +1,5 @@
 [toolchain]
-channel    = "nightly-2022-02-16"
+channel    = "nightly-2022-02-02"
 components = [ "clippy", "rustfmt" ]
 profile    = "minimal"
 targets    = [ "wasm32-unknown-unknown" ]
